Update photo handling in Participation service
The photo handling in the Participation service has been improved with two changes: the ability to add and receive an image for a Participation entity. Adjustments have been made so that the upload and download processes now work via API calls. Furthermore, the application.yml has been updated to limit the size of uploaded files to 5MB.