Add satisfaction score and comment to Participation
A new SatisfactionForm class has been created to handle the addition of satisfaction scores and comments to a Participation instance. The ParticipationService interface has been updated to include an addSatisfaction method, and this method has been implemented in the ParticipationServiceImpl class. A unit test for this new functionality has also been added.