Add rate limiting via Bucket4j
This update introduces rate limiting functionality to the application. A new `RateLimitInterceptor` was created and registered in the `WebMvcConfig`. If a limit is exceeded, a `TooManyRequestsException` is now thrown and handled appropriately in `ControllerAdvisor`. Bucket4j is used for rate limiting, and therefore its dependency was added to the pom.xml.